pyproject.toml
[tool.poetry]
name = "innoter-drf"
version = "0.1.0"
description = ""
authors = ["arti <loscoarti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
black = "^23.11.0"
isort = "^5.13.0"
python-dotenv = "^1.0.0"
pre-commit = "^3.6.0"
django-filter = "^23.5"
pyjwt = "^2.8.0"
celery = "^5.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
